+++ common/src/main/java/com/reazip/economycraft/shop/ShopSearchable.java
@@ -1,5 +1,0 @@
-package com.reazip.economycraft.shop;
-
-public interface ShopSearchable {
-    void applySearch(String query);
-}